+++ go/testdata/pan-os_simul.t
@@ -1,5 +1,5 @@
 =TEMPL=apikey
-GET /api?type=keygen
+POST /api/?type=keygen
 <response status = 'success'>
  <result><key>LUFRPT=</key></result>
 </response>
@@ -7,7 +7,7 @@ GET /api?type=keygen
 
 =TEMPL=checkHA
 [[apikey]]
-GET /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
+POST /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
 <response status = 'success'>
  <result>
   <enabled>{{or .enabled "yes"}}</enabled>
@@ -23,7 +23,7 @@ GET /api/?type=op&cmd=<show><high-availability><state/></high-availability></sho
 =END=
 
 =TEMPL=empty_missing_vsys
-GET /api/?type=config&action=get&xpath=/config/devices
+POST /api/?type=config&action=get&xpath=/config/devices
 <response status = 'success'>
  <result>
   <devices>
@@ -41,7 +41,7 @@ GET /api/?type=config&action=get&xpath=/config/devices
 
 =TEMPL=empty_with_vsys
 [[checkHA]]
-GET /api/?type=config&action=get&xpath=/config/devices
+POST /api/?type=config&action=get&xpath=/config/devices
 <response status = 'success'>
  <result>
   <devices>
@@ -65,7 +65,7 @@ GET /api/?type=config&action=get&xpath=/config/devices
 ############################################################
 =TITLE=Device gives status 500
 =SCENARIO=
-GET /api
+POST /api/
 500
 device not ready
 =NETSPOC=NONE
@@ -75,7 +75,8 @@ WARNING>>> device not ready
 ERROR>>> Devices unreachable: router
 =OUTPUT=
 --router.login
-TESTSERVER/api?password=xxx&type=keygen&user=admin
+TESTSERVER/api/
+DATA: password=xxx&type=keygen&user=admin
 device not ready
 
 =END=
@@ -83,22 +84,23 @@ device not ready
 ############################################################
 =TITLE=Device gives no valid answer
 =SCENARIO=
-GET /api
+POST /api/
 EOF
 =NETSPOC=NONE
 =ERROR=
-WARNING>>> API key Get "TESTSERVER/api?password=xxx&type=keygen&user=admin": EOF
+WARNING>>> API key Post "TESTSERVER/api/": EOF
 ERROR>>> Devices unreachable: router
 =OUTPUT=
 --router.login
-TESTSERVER/api?password=xxx&type=keygen&user=admin
+TESTSERVER/api/
+DATA: password=xxx&type=keygen&user=admin
 
 =END=
 
 ############################################################
 =TITLE=Login fails
 =SCENARIO=
-GET /api?type=keygen
+POST /api/?type=keygen
 <response status = 'failure'>
  <msg>User unknown</msg>
 </response>
@@ -118,12 +120,14 @@ WARNING>>> not in active state: 10.1.13.33 (router)
 ERROR>>> Devices unreachable: router
 =OUTPUT=
 --router.login
-TESTSERVER/api?password=xxx&type=keygen&user=admin
+TESTSERVER/api/
+DATA: password=xxx&type=keygen&user=admin
 <response status = 'success'>
  <result><key>xxx</key></result>
 </response>
 
-TESTSERVER/api/?key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
+TESTSERVER/api/
+DATA: key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
 404 page not found
 
 =END=
@@ -132,7 +136,7 @@ TESTSERVER/api/?key=xxx&type=op&cmd=<show><high-availability><state/></high-avai
 =TITLE=HA check fails with EOF
 =SCENARIO=
 [[apikey]]
-GET /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
+POST /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
 <INVALID/>
 =NETSPOC=NONE
 =ERROR=
@@ -144,7 +148,7 @@ ERROR>>> Devices unreachable: router
 =TITLE=HA check can't parse HA mode
 =SCENARIO=
 [[apikey]]
-GET /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
+POST /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
 <response status = 'success'>
  <INVALID/>
 </response>
@@ -164,12 +168,14 @@ ERROR>>> status code: 404
 ERROR>>> 404 page not found
 =OUTPUT=
 --router.login
-TESTSERVER/api?password=xxx&type=keygen&user=admin
+TESTSERVER/api/
+DATA: password=xxx&type=keygen&user=admin
 <response status = 'success'>
  <result><key>xxx</key></result>
 </response>
 
-TESTSERVER/api/?key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
+TESTSERVER/api/
+DATA: key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
 <response status = 'success'>
  <result>
   <enabled>yes</enabled>
@@ -184,7 +190,8 @@ TESTSERVER/api/?key=xxx&type=op&cmd=<show><high-availability><state/></high-avai
 </response>
 
 --router.config
-TESTSERVER/api/?key=xxx&type=config&action=get&xpath=/config/devices
+TESTSERVER/api/
+DATA: key=xxx&type=config&action=get&xpath=/config/devices
 404 page not found
 
 =END=
@@ -193,7 +200,7 @@ TESTSERVER/api/?key=xxx&type=config&action=get&xpath=/config/devices
 =TITLE=HA not enabled
 =SCENARIO=
 [[apikey]]
-GET /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
+POST /api/?type=op&cmd=<show><high-availability><state/></high-availability></show>
 <response status = 'success'>
  <result>
   <enabled>no</enabled>
@@ -203,12 +210,14 @@ GET /api/?type=op&cmd=<show><high-availability><state/></high-availability></sho
 =NETSPOC=NONE
 =OUTPUT=
 --router.login
-TESTSERVER/api?password=xxx&type=keygen&user=admin
+TESTSERVER/api/
+DATA: password=xxx&type=keygen&user=admin
 <response status = 'success'>
  <result><key>xxx</key></result>
 </response>
 
-TESTSERVER/api/?key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
+TESTSERVER/api/
+DATA: key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
 <response status = 'success'>
  <result>
   <enabled>no</enabled>
@@ -228,12 +237,14 @@ state: active-primary
 =NETSPOC=NONE
 =OUTPUT=
 --router.login
-TESTSERVER/api?password=xxx&type=keygen&user=admin
+TESTSERVER/api/
+DATA: password=xxx&type=keygen&user=admin
 <response status = 'success'>
  <result><key>xxx</key></result>
 </response>
 
-TESTSERVER/api/?key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
+TESTSERVER/api/
+DATA: key=xxx&type=op&cmd=<show><high-availability><state/></high-availability></show>
 <response status = 'success'>
  <result>
   <enabled>yes</enabled>
@@ -266,7 +277,7 @@ ERROR>>> Devices unreachable: router
 =TITLE=Invalid config in response
 =SCENARIO=
 [[checkHA]]
-GET /api/?type=config&action=get&xpath=/config/devices
+POST /api/?type=config&action=get&xpath=/config/devices
 <INVALID>
 =NETSPOC=NONE
 =ERROR=
@@ -277,7 +288,7 @@ ERROR>>> While reading device: Parsing response: expected element type <response
 =TITLE=Invalid config in result
 =SCENARIO=
 [[checkHA]]
-GET /api/?type=config&action=get&xpath=/config/devices
+POST /api/?type=config&action=get&xpath=/config/devices
 <response status = 'success'>
  <result>
   <devices>
@@ -294,7 +305,7 @@ ERROR>>> While reading device: Parsing response: XML syntax error on line 4: ele
 =TITLE=Empty config with missing device name
 =SCENARIO=
 [[checkHA]]
-GET /api/?type=config&action=get&xpath=/config/devices
+POST /api/?type=config&action=get&xpath=/config/devices
 <response status = 'success'>
  <result>
   <devices>
@@ -355,35 +366,35 @@ ERROR>>> Unknown name 'vsys1' in VSYS of device configuration
 =TITLE=Changing device fails
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 EOF
 =NETSPOC=
 [[minimal_netspoc]]
 =ERROR=
-ERROR>>> Command failed with Get "TESTSERVER/api/?key=LUFRPT=&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/service/entry[@name='tcp+80']&element=%3Cprotocol%3E%3Ctcp%3E%3Cport%3E80%3C%2Fport%3E%3C%2Ftcp%3E%3C%2Fprotocol%3E": EOF
+ERROR>>> Command failed with Post "TESTSERVER/api/": EOF
 =END=
 
 ############################################################
 =TITLE=Committing changes fails
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 EOF
 =NETSPOC=
 [[minimal_netspoc]]
 =ERROR=
-ERROR>>> Commit failed: Get "TESTSERVER/api/?key=LUFRPT=&type=commit&action=partial&cmd=<commit><partial><admin><member>admin</member></admin></partial></commit>": EOF
+ERROR>>> Commit failed: Post "TESTSERVER/api/": EOF
 =END=
 
 ############################################################
 =TITLE=Invalid XML in answer from commit
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 <invalid/>
 =NETSPOC=
 [[minimal_netspoc]]
@@ -395,9 +406,9 @@ ERROR>>> Commit failed: Parsing response: expected element type <response> but h
 =TITLE=Useless commit
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 <response status="success">
  <msg>The result of this commit would be the same</msg>
 </response>
@@ -405,13 +416,16 @@ GET /api/?type=commit&action=partial
 [[minimal_netspoc]]
 =OUTPUT=
 --router.change
-TESTSERVER/api/?key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/service/entry[@name='tcp 80']&element=<protocol><tcp><port>80</port></tcp></protocol>
+TESTSERVER/api/
+DATA: key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/service/entry[@name='tcp 80']&element=<protocol><tcp><port>80</port></tcp></protocol>
 <response status="success" code="20"></response>
 
-TESTSERVER/api/?key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/rulebase/security/rules/entry[@name='r1']&element=<action>allow</action><from><member>z1</member></from><to><member>z2</member></to><source><member>any</member></source><destination><member>any</member></destination><service><member>tcp 80</member></service><application><member>any</member></application><rule-type>interzone</rule-type>
+TESTSERVER/api/
+DATA: key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/rulebase/security/rules/entry[@name='r1']&element=<action>allow</action><from><member>z1</member></from><to><member>z2</member></to><source><member>any</member></source><destination><member>any</member></destination><service><member>tcp 80</member></service><application><member>any</member></application><rule-type>interzone</rule-type>
 <response status="success" code="20"></response>
 
-TESTSERVER/api/?key=xxx&type=commit&action=partial&cmd=<commit><partial><admin><member>admin</member></admin></partial></commit>
+TESTSERVER/api/
+DATA: key=xxx&type=commit&action=partial&cmd=<commit><partial><admin><member>admin</member></admin></partial></commit>
 <response status="success">
  <msg>The result of this commit would be the same</msg>
 </response>
@@ -422,9 +436,9 @@ TESTSERVER/api/?key=xxx&type=commit&action=partial&cmd=<commit><partial><admin><
 =TITLE=Commit with unknown message
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 <response status="success"><msg>Unknown</msg></response>
 =NETSPOC=
 [[minimal_netspoc]]
@@ -436,27 +450,27 @@ ERROR>>> Commit failed: Unexpected message: Unknown
 =TITLE=Getting job status fails
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 <response status="success" code="19"><result><job>6</job></result></response>
-GET /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
+POST /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
 EOF
 =NETSPOC=
 [[minimal_netspoc]]
 =ERROR=
-ERROR>>> Commit failed: Get "TESTSERVER/api/?key=LUFRPT=&type=op&cmd=<show><jobs><id>6</id></jobs></show>": EOF
+ERROR>>> Commit failed: Post "TESTSERVER/api/": EOF
 =END=
 
 ############################################################
 =TITLE=Invalid response in job status
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 <response status="success" code="19"><result><job>6</job></result></response>
-GET /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
+POST /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
 <invalid/>
 =NETSPOC=
 [[minimal_netspoc]]
@@ -468,11 +482,11 @@ ERROR>>> Commit failed: Parsing response: expected element type <response> but h
 =TITLE=Unexpected message in job status
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 <response status="success" code="19"><result><job>6</job></result></response>
-GET /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
+POST /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
 <response status="success"><result><job>
 <result>invalid</result>
 </job></result></response>
@@ -486,11 +500,11 @@ ERROR>>> Commit failed: Unexpected job result: "invalid"
 =TITLE=Add rule and service to empty device, successful commit
 =SCENARIO=
 [[empty_with_vsys]]
-GET /api/?action=set&type=config
+POST /api/?action=set&type=config
 <response status="success" code="20"></response>
-GET /api/?type=commit&action=partial
+POST /api/?type=commit&action=partial
 <response status="success" code="19"><result><job>6</job></result></response>
-GET /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
+POST /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
 <response status="success"><result><job>
 <result>OK</result>
 </job></result></response>
@@ -498,16 +512,20 @@ GET /api/?type=op&cmd=<show><jobs><id>6</id></jobs></show>
 [[minimal_netspoc]]
 =OUTPUT=
 --router.change
-TESTSERVER/api/?key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/service/entry[@name='tcp 80']&element=<protocol><tcp><port>80</port></tcp></protocol>
+TESTSERVER/api/
+DATA: key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/service/entry[@name='tcp 80']&element=<protocol><tcp><port>80</port></tcp></protocol>
 <response status="success" code="20"></response>
 
-TESTSERVER/api/?key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/rulebase/security/rules/entry[@name='r1']&element=<action>allow</action><from><member>z1</member></from><to><member>z2</member></to><source><member>any</member></source><destination><member>any</member></destination><service><member>tcp 80</member></service><application><member>any</member></application><rule-type>interzone</rule-type>
+TESTSERVER/api/
+DATA: key=xxx&action=set&type=config&xpath=/config/devices/entry[@name='localhost.localdomain']/vsys/entry[@name='vsys1']/rulebase/security/rules/entry[@name='r1']&element=<action>allow</action><from><member>z1</member></from><to><member>z2</member></to><source><member>any</member></source><destination><member>any</member></destination><service><member>tcp 80</member></service><application><member>any</member></application><rule-type>interzone</rule-type>
 <response status="success" code="20"></response>
 
-TESTSERVER/api/?key=xxx&type=commit&action=partial&cmd=<commit><partial><admin><member>admin</member></admin></partial></commit>
+TESTSERVER/api/
+DATA: key=xxx&type=commit&action=partial&cmd=<commit><partial><admin><member>admin</member></admin></partial></commit>
 <response status="success" code="19"><result><job>6</job></result></response>
 
-TESTSERVER/api/?key=xxx&type=op&cmd=<show><jobs><id>6</id></jobs></show>
+TESTSERVER/api/
+DATA: key=xxx&type=op&cmd=<show><jobs><id>6</id></jobs></show>
 <response status="success"><result><job>
 <result>OK</result>
 </job></result></response>